requirements.txt
Django==5.0.7
python-dotenv==1.0.1
str2bool==1.1
django-allauth
whitenoise==6.5.0
django-debug-toolbar==4.2.0

mssql-django

celery==5.3.4
redis==5.0.1
hiredis==2.2.3

django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
drf-spectacular==0.26.5

Pillow==10.0.1

sentry-sdk[django]

kubernetes==30.1.0
httpx~=0.27.0

channels[daphne]~=4.1.0
Twisted[tls,http2]
daphne
authlib~=1.3.1
websockets~=12.0

asgiref~=3.8.1
PyYAML~=6.0.1
urllib3~=2.2.2
pytz~=2024.1

boto3~=1.34.147
algoliasearch-django>=3.0,<4.0

botocore~=1.34.147